requirements.txt
-i https://pypi.org/simple
alembic==1.9.2; python_version >= '3.7'
click==8.1.3; python_version >= '3.7'
flask==2.2.2; python_version >= '3.7'
flask-cors==3.0.10
flask-login==0.6.2; python_version >= '3.7'
flask-migrate==4.0.2; python_version >= '3.6'
flask-sqlalchemy==3.0.2; python_version >= '3.7'
flask-wtf==1.1.1; python_version >= '3.7'
greenlet==3.0.1; python_version >= '3.7'
gunicorn==20.1.0; python_version >= '3.5'
importlib-metadata==6.9.0; python_version < '3.10'
itsdangerous==2.1.2; python_version >= '3.7'
jinja2==3.1.2; python_version >= '3.7'
mako==1.2.4; python_version >= '3.7'
markupsafe==2.1.2; python_version >= '3.7'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
python-dotenv==1.0.0
python-editor==1.0.4
setuptools==69.0.2; python_version >= '3.8'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
sqlalchemy==1.4.46; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
werkzeug==2.2.2; python_version >= '3.7'
wtforms==3.0.1; python_version >= '3.7'
zipp==3.17.0; python_version >= '3.8'
alpaca-py==0.13.3
aiohttp==3.8.5
flask-socketio==5.3.4
gevent-websocket==0.10.1
PyJWT==2.8.0
websockets==12.0
pytest-asyncio==0.23.5
pytz==2024.1
